Rework exit status and add a summary on error
Previously the exit status could be an i32, which really doesn't make sense on
Unix. Now it will either be 0 or 1.

In addition, I also added a summary output when there are (expected) errors
from tidying or linting. When precious exits it shows you all failures, one
per filter + path combo. This is primarily useful for linting, as tidying
failures typically occur for non-recoverable reasons (misconfiguration of the
command, the external command is missing, etc.).

+++ src/main.rs
@@ -29,5 +29,5 @@ fn main() {
             127
         }
     };
-    std::process::exit(status);
+    std::process::exit(status as i32);
 }
